% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_count.R
\name{dsi_count}
\alias{dsi_count}
\title{Obtain a count of country records from an NCBI database}
\usage{
dsi_count(country = NULL, db = "nuccore")
}
\arguments{
\item{country}{A country name or country and date or other terms (see
examples)}

\item{db}{database to search}
}
\value{
character
}
\description{
Search NCBI entrez databases to create counts of country information.

This is a wrapper around \code{rentrexz:rentrez_search()} from the \code{rentrez
  \url{https://github.com/ropensci/rentrez}} package.

 For a list of entrez databases use rentrez::entrez_dbs().
}
\details{
For available fields for query construction see
  \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/}. For functions in the
  \code{rentrez} package see \url{https://github.com/ropensci/rentrez}
}
\examples{
bahamas <- dsi_count(country = "bahamas", db = "nuccore")
dsi_count(country = "bahamas", db = "biosample")
dsi_count(country = "bahamas", db = "bioproject")
dsi_count(country = "bahamas", db = "protein")

# Get counts with a publication date
kenya_pub <- dsi_count(country = "(Kenya) AND 2020/02/26[PDAT]", db = "nuccore", )
kenya_pub_range <- dsi_count(country = "(Kenya) AND 2020/01/01:2020/05/18[PDAT]", db = "nuccore")

# Get counts with a modification date
kenya_mod <- dsi_count(country = "(Kenya) AND 2020/05/18[Modification Date]", db = "nuccore", )
kenya_mod_range <- dsi_count(term = "(Kenya) AND 2020/01/01:2020/05/18[Modification Date]", db = "nuccore")

# Get counts for and without Homo sapiens and check they add up
kenya_all <- dsi_count(country = "kenya", db = "nuccore") # 469351
kenya_homo <- dsi_count(country = "(Kenya AND Homo sapiens[ORGN])", db = "nuccore") # 2313
kenya_not <- dsi_count(country = "(Kenya NOT Homo sapiens[ORGN])", db = "nuccore") # 467038
(kenya_homo + kenya_not) == kenya_all # expect TRUE
}
