% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_country.R
\name{dsi_country}
\alias{dsi_country}
\title{Search NCBI entrez databases for country information}
\usage{
dsi_country(country = NULL, db = "nuccore")
}
\arguments{
\item{country}{A country name or country and date or other terms (see
examples)}

\item{db}{database to search}
}
\value{
a list of xml objects
}
\description{
This function is for searching NCBI entrez databases for country information.
The function is a simplified wrapper for the \code{rentrez::entrez_search()} function
\url{https://github.com/ropensci/rentrez}. For a list
of entrez databases use \code{rentrez::entrez_dbs()} The function is designed to
download the summary data for the results of the search which can be many
Gibabytes (e.g. 2.4 million records from South Africa is 45Gb). The function
does not download the sequences. See the rentrez package to download the
actual sequences in different formats.
}
\details{
For available fields for query construction see
\url{https://www.ncbi.nlm.nih.gov/books/NBK49540/}. For functions in the
\code{rentrez} package see \url{https://github.com/ropensci/rentrez}
}
\examples{
\dontrun{bahamas <- dsi_country(country = "bahamas", db = "nuccore")

# Get records with a publication date kenya_pub <- dsi_country(country =
"(Kenya) AND 2020/02/26[PDAT]", db = "nuccore", use_history = TRUE)
kenya_pub_range <- dsi_country(country = "(Kenya) AND
2020/01/01:2020/05/18[PDAT]", db = "nuccore", use_history = TRUE)

# Get records with a modification date kenya_mod <- dsi_country(country =
"(Kenya) AND 2020/05/18[Modification Date]", db = "nuccore",  use_history =
TRUE) kenya_mod_range <- dsi_country(country = "(Kenya) AND
2020/01/01:2020/05/18[Modification Date]", db = "nuccore",  use_history =
TRUE) }
}
