% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_parse.R
\name{dsi_parse}
\alias{dsi_parse}
\title{Parse the Esummary from rentrez_country to a data frame}
\usage{
dsi_parse(x)
}
\arguments{
\item{x}{a list object from entrez_summary nucleotide database}
}
\value{
data frame
}
\description{
Function to parse the results of a call to rentrez_fetch to a data frame. Note that does not presently include the statistics
Tested on 10,0000 records for the Bahamas
}
\details{
This function returns all available esummary meta data except the statistics chunk (as this partly repeats information such as sequence length)
}
\examples{
\dontrun{
# get the data
bahamas_country <- dsi_country(country = "bahamas", dbase = "nuccore")
# map it to a data frame
bahamas_sequences <- map_df(bahamas_country, dsi_parse)
}
}
