% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pic_cty}
\alias{pic_cty}
\title{Pacific Island Country ISO3 codes}
\format{
A character vector of 14 ISO 3166-1 alpha-3 codes.
}
\usage{
pic_cty
}
\description{
Character vector of ISO3 codes for the 14 WHO Member States classified
as Pacific Island Countries (PICs). Sorted alphabetically.
}
\details{
All 14 PICs are within the Western Pacific Region, so
\code{all(pic_cty \%in\% wpro_cty)} is \code{TRUE}. Non-Member Pacific areas
(e.g. New Caledonia, French Polynesia, American Samoa, Tokelau) are
not included.

The 14 PIC Member States are: Cook Islands, Fiji, Kiribati, Marshall
Islands, Micronesia (Federated States of), Nauru, Niue, Palau, Papua
New Guinea, Samoa, Solomon Islands, Tonga, Tuvalu, and Vanuatu.
}
\examples{
length(pic_cty)              # 14
all(pic_cty \%in\% wpro_cty)   # TRUE
}
\seealso{
\code{\link{who_countries}}, \code{\link{wpro_cty}}.
}
\keyword{datasets}
