% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentrez_country.R
\name{rentrez_country}
\alias{rentrez_country}
\title{Search NCBI entrez databases for country information
Use the rentrez package to search an NCBI Entrez database for country information. For a list of entrez databases use rentrez::entrez_dbs()
The function is designed to download the summary data for the results of the search which can be many Gibabytes (e.g. 2.4 million records from South Africa is 45Gb).
The function does not download the sequences. See the rentrez package to download the actual sequences in different formats.}
\usage{
rentrez_country(country = NULL, dbase = "nucleotide", date = NULL)
}
\arguments{
\item{country}{}

\item{db}{}
}
\value{

}
\description{
Search NCBI entrez databases for country information
Use the rentrez package to search an NCBI Entrez database for country information. For a list of entrez databases use rentrez::entrez_dbs()
The function is designed to download the summary data for the results of the search which can be many Gibabytes (e.g. 2.4 million records from South Africa is 45Gb).
The function does not download the sequences. See the rentrez package to download the actual sequences in different formats.
}
\details{
For available fields for query constructions see \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/}
}
\examples{
bahamas <- rentrez_country(country = "bahamas", dbase = "nucleotide")

# Get records with a publication date
kenya_pub <- rentrez::entrez_search(db = "nucleotide", term = "(Kenya) AND 2020/02/26[PDAT]", use_history = TRUE)
kenya_pub_range <- rentrez::entrez_search(db = "nucleotide", term = "(Kenya) AND 2020/01/01:2020/05/18[PDAT]", use_history = TRUE)

# Get records with a modification date
kenya_mod <- rentrez::entrez_search(db = "nucleotide", term = "(Kenya) AND 2020/05/18[Modification Date]", use_history = TRUE)
kenya_mod_range <- rentrez::entrez_search(db = "nucleotide", term = "(Kenya) AND 2020/01/01:2020/05/18[Modification Date]", use_history = TRUE)
}
