% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentrez_parse2.R
\name{rentrez_parse2}
\alias{rentrez_parse2}
\title{Parse Esummary from rentrez to data frame
Function to parse the results of a call to rentrez_fetch to a data frame. Note that does not presently include the statistics
Tested on 10,0000 records for the Bahamas}
\usage{
rentrez_parse2(x)
}
\arguments{
\item{x}{a list object from entrez_summary nucleotide database}
}
\value{
data frame
}
\description{
Parse Esummary from rentrez to data frame
Function to parse the results of a call to rentrez_fetch to a data frame. Note that does not presently include the statistics
Tested on 10,0000 records for the Bahamas
}
\examples{
\dontrun{

# with a single result set

bs_nucleotide <- entrez_search(db = "nucleotide",
term = "Bahamas", use_history = TRUE), then 
bs_summary <- entrez_summary(db="nucleotide", web_history = bs_nucleotide$web_history, 
retmode="xml", retmax=10000, retstart=seq_start) then
bs_results <- entrez_parse(bs_summary)

# With multiple list objects

bahamas_country <- rentrez_country(country = "bahamas", dbase = "nucleotide")
bahamas_sequences <- map_df(bahamas_sequences, rentrez_parse)

}
}
