% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpro_cty}
\alias{wpro_cty}
\title{WPRO Member States (ISO 3166-1 alpha-3)}
\format{
A character vector of length N.
}
\usage{
wpro_cty
}
\description{
A character vector of ISO3 codes for countries and areas
in the WHO Western Pacific Region.
}
\keyword{datasets}
